player prefab fixed
fixed issue with extra camera on a 3d model. created a separate prefab for the player, please use that one. it can be very much also used for functionality and mechanics stuff just make a parent object preferably that controls the model as i'm likely to change the base when we get to animation phase after art prototype