Adding Luna Materials/Model & Fixing Up Church
Set up the materials for Luna and added them to the animated prefab. Separated the animation from that prefab so it can be used more easily. Added a non animated Luna model. Imported a new version of the church model and added the material to it since the initial asymmetrical model was causing noticeable distortion to the texture.